+++ src/lib.rs
@@ -299,7 +299,7 @@ pub use crate::notifications::{
 
 pub use crate::state::{
     Chain, ChainParams, ChainType, ChainstateManager, ChainstateManagerBuilder, Context,
-    ContextBuilder, ProcessBlockHeaderResult, ProcessBlockResult,
+    ContextBuilder, ProcessBlockHeaderResult, ProcessBlockResult, ValidateBlockResult,
 };
 
 pub use crate::core::block_check_flags::{