+++ src/state/mod.rs
@@ -5,5 +5,6 @@ pub mod context;
 pub use chain::{Chain, ChainIterator};
 pub use chainstate::{
     ChainstateManager, ChainstateManagerBuilder, ProcessBlockHeaderResult, ProcessBlockResult,
+    ValidateBlockResult,
 };
 pub use context::{ChainParams, ChainType, Context, ContextBuilder};